==> 2_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import string
import nltk
import re
nltk.download('punkt')
nltk.download('stopwords')
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
import datetime

<p style='font-size:20px'><b> Data preprocessing </b></p>
<p>

- The raw data available in the 'Title' & 'Body' columns is collection of text, paragraphs, code, links, etc that can't be directly used as input data for training.
- As part of data preprocessing, we will derive individual words/phrases from these documents that can be used as features for training the model.
- Considering the nature of the data, we can assume that most of the content in Title & Body paragraphs will have words to build tag association  

Following NLP preprocessing steps were applied to achieve the desired features/words in the Title & Body Paragraphs:

    1. Punctuation removal
    2. Lower case conversion
    3. Stopword removal
    4. Stemming

These steps are first applied to the title only to see the before & after at each step. Later it has been applied to the texts, paragraphs filtered in the body. 

The code segments in the Body contain important information in symbols like '#' , ';' & therefore we won't apply any preprocessing steps, but rather directly derive char n-grams from complete code.   
    
</p>

start_time = datetime.datetime.now()

# Load the trimmed data set of 1.25L datapoints for preprocessing
raw_data = pd.read_csv('E:/ai_projects/data_storage/Stack over flow tagging 5/dataset/trimmed.csv')

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Step1: Punctuation removal

def punct_removal(s):

    # Replacing all punctuation marks with space character
    punctuation_free = s.translate(str.maketrans(string.punctuation, len(string.punctuation)*' ', ''))
    return punctuation_free

raw_data['Title_punctfree'] = raw_data.Title.apply(punct_removal)

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_punctfree', 'Body', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before punctuation free: {}".format(i+1, raw_data.Title[random_indices[i]]))
    print("{}b. After punctuation free: {}\n".format(i+1, raw_data.Title_punctfree[random_indices[i]]))
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Step2: Converting to lower case
raw_data['Title_lower'] = raw_data.Title_punctfree.apply(lambda s:s.lower())

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_punctfree', 'Title_lower', 'Body', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before lower case: {}".format(i+1, raw_data.Title_punctfree[random_indices[i]]))
    print("{}b. After lower case: {}\n".format(i+1, raw_data.Title_lower[random_indices[i]]))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Tokenizing the words
raw_data['Title_tokens'] = raw_data.Title_lower.apply(lambda s:nltk.tokenize.word_tokenize(s))

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_punctfree', 'Title_lower', 'Title_tokens', 'Body', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before tokenizing: {}".format(i+1, raw_data.Title_lower[random_indices[i]]))
    print("{}b. After tokenizing: {}\n".format(i+1, raw_data.Title_tokens[random_indices[i]]))
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Step3: Stopword removal

# Carefully choose which stopwords to remove based on problem, e.g 'not' keyword would be important in some NLP modelling problems like sentiment detection
stopwords  = nltk.corpus.stopwords.words('english')

def stopword_removal(tokens):
    global stopwords
    filtered_tokens = [item for item in tokens if item not in stopwords]
    return filtered_tokens

raw_data['Title_stop_removed'] = raw_data.Title_tokens.apply(stopword_removal)

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_punctfree', 'Title_lower', 'Title_tokens', 'Title_stop_removed', 'Body', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before stopword removal: {}".format(i+1, raw_data.Title_lower[random_indices[i]]))
    print("{}b. After stopword removal: {}".format(i+1, raw_data.Title_stop_removed[random_indices[i]]))
    print("{}c. Tags: {}\n".format(i+1, raw_data.Tags_shortlist[random_indices[i]]))
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Step4: Stemming

# Try other stemming algorithms like snowball, lancaster, etc & understand what is the difference?
porter_stem = nltk.stem.PorterStemmer()

def stem_token(tokens):
    
    global porter_stem
    stemmed_tokens = [porter_stem.stem(item) for item in tokens]
    return stemmed_tokens

raw_data['Title_stemmed'] = raw_data.Title_stop_removed.apply(stem_token)

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_lower', 'Title_stemmed', 'Body', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before stemming: {}".format(i+1, raw_data.Title_lower[random_indices[i]]))
    print("{}b. After stemming: {}\n".format(i+1, raw_data.Title_stemmed[random_indices[i]]))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

# Since title are not enough take a look at body of few examples

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Title: {}".format(i+1, raw_data.Title[random_indices[i]]))
    print("{}b. Body: {}".format(i+1, raw_data.Body[random_indices[i]]))
    print("{}c. Tags: {}\n\n\n".format(i+1, raw_data.Tags_shortlist[random_indices[i]]))


start_time = datetime.datetime.now()

def body_filter(s):

    # Separate the code blocks from the body
    code_regex = r'<code>(.*?)</code>'
    code_blocks = re.findall(code_regex, s, flags=re.MULTILINE|re.DOTALL)
    s = re.sub(code_regex, ' ', s, flags=re.MULTILINE|re.DOTALL)

    # Remove html tags from the body
    # Note this will also remove hyper links (<a href="link">), which may contain valuable information
    html_regex = r'<(.*?)>'
    s = re.sub(html_regex, ' ', s, flags=re.MULTILINE|re.DOTALL)
    
    return [s, code_blocks]
    
    
temp = raw_data.Body.apply(body_filter)

raw_data['Body_text'] = [temp[i][0] for i in range(len(temp))]
raw_data['Body_code'] = [temp[i][1] for i in range(len(temp))]

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_stemmed', 'Body', 'Body_text', 'Body_code', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("-------- Example {} --------\n".format(i+1))
    print("{}a. Body:\n {}".format(i+1, raw_data.Body[random_indices[i]]))
    print("\n\n{}b. Text:\n {}".format(i+1, raw_data.Body_text[random_indices[i]]))
    print("\n\n{}c. Code:\n {}\n\n\n\n".format(i+1, raw_data.Body_code[random_indices[i]]))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Merging all fragmented code blocks in a datapoint

def merge_code(code_blocks):
    return " " + " ".join(code_blocks)

raw_data['Code'] = raw_data.Body_code.apply(merge_code)

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_stemmed', 'Body', 'Body_text', 'Body_code', 'Code', 'Tags_shortlist']]

print("Few examples:\n")
random_indices = np.random.randint(0, len(raw_data.Title), 5)

for i in range(5):
    print("{}a. Before: {}".format(i+1, raw_data.Body_code[random_indices[i]]))
    print("{}b. After: {}\n".format(i+1, raw_data.Code[random_indices[i]]))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Apply all the preprocessing steps to the body as applied to title

def body_preprocessing(s):

    # Step1: Punctuation removal
    punctuation_free = s.translate(str.maketrans('', '', string.punctuation))

    # Step2: Converting to lower case
    s_lower  = punctuation_free.lower()

    # Tokenizing the sentence
    tokens = nltk.tokenize.word_tokenize(s_lower)

    # Step3: Stopword removal
    global stopwords
    filtered_tokens = [item for item in tokens if item not in stopwords]

    # Step4: Stemming
    global porter_stem
    stemmed_tokens = [porter_stem.stem(item) for item in filtered_tokens]
    
    return stemmed_tokens


raw_data['Body_stemmed'] = raw_data.Body_text.apply(body_preprocessing)

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_stemmed', 'Body', 'Body_stemmed', 'Code', 'Tags_shortlist']]

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Merging the processed tokens to save the datapoint as a single raw string
# To provide more weightage to title since it would contain more relevant & crisp information, the final raw string has title added 3 times vs body tokens added once 

raw_data['Title_processed'] = [" ".join(item) for item in raw_data.Title_stemmed]
raw_data['Body_processed'] = [" ".join(item) for item in raw_data.Body_stemmed]
raw_data['Title3_Body'] = [raw_data.Title_processed[i] + " " + raw_data.Title_processed[i] + " " + raw_data.Title_processed[i] + " " + raw_data.Body_processed[i] for i in range(raw_data.shape[0])]

# Rearranging the dataset
raw_data = raw_data.loc[:, ['Id', 'Title', 'Title_processed', 'Body', 'Body_processed', 'Title3_Body', 'Code', 'Tags_shortlist']]

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Save data to csv file
#### Imp: Careful not to overwrite data ####

try:
    raw_data.to_csv('dataset/processed.csv', mode = 'x', index = False)
except:
    print("File already saved?")
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)    

start_time = datetime.datetime.now()

# Train CV test split: Split the data in training, CV and test

temp_data, test_data = train_test_split(raw_data, train_size = 125000)
train_data, cv_data = train_test_split(temp_data, train_size = 100000)

print(train_data.shape)
print(cv_data.shape)
print(test_data.shape)

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

try:
    train_data.to_csv('dataset/processed_train.csv', mode = 'x', index = False)
except:
    print("File already saved?")

try:
    cv_data.to_csv('dataset/processed_cv.csv', mode = 'x', index = False)
except:
    print("File already saved?")
    
    
try:
    test_data.to_csv('dataset/processed_test.csv', mode = 'x', index = False)
except:
    print("File already saved?")

print("\nBlock execution time: ", datetime.datetime.now() - start_time)
