==> 5_reiterate.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from operator import itemgetter
from sklearn.feature_extraction.text import CountVectorizer
import random
import datetime

<p style='font-size:20px'><b> Dataset re-evaluation </b></p>
<p>

Based on the modelling implemented, following are the best result obtained on test data so far: 

Model A) Log-reg, l1 penalty, alpha = 3.25e-6 | Micro-avg F1-score = 56.0%, Macro-avg F1-score = 43.2%

Model B) Lin-SVM, l1 penalty, alpha = 3.25e-6 | Micro-avg F1-score = 57.1%, Macro-avg F1-score = 42.1%

- In both cases, the Macro-avg score is quite low compared to the Micro-avg F1 score.
- This implies that the F1-score for the less frequent tags is quite lower that the top tags which may be due low datapoints having these tags since the points were sampled randomply.
- To rectify this, we will now sample the data to capture atleast 1000 datapoints for each of the 500 tags to be predicted.
    
</p>


start_time = datetime.datetime.now()

# Load the Train.csv dataset for Stack over flow tagging problem 
raw_data = pd.read_csv('dataset/Train.csv')

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

<p style='font-size:16px'><b> Old dataset comparison (1.5L points)  </b></p>

<p>

Plotting the number of data points associated to each tag in the old dataset havng 1.5L datapoints. 
    
</p>


start_time = datetime.datetime.now()

# Load the trimmed data set of 1.25L datapoints for preprocessing
raw_data_old = pd.read_csv('dataset/trimmed.csv')

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# List to store each tagging instance accross all datapoints
tag_list = []

tags = [str(x).split(" ") for x in raw_data_old.Tags_shortlist]

for datapoint in tags:
    for tag in datapoint:
        tag_list.append(tag)

# Calculating the unique tags & their frequency in a descending order from the above tag_list        
unique_tags, tag_count = np.unique(tag_list, return_counts=True)
sort_indices = np.argsort(tag_count)[::-1]
unique_tags = unique_tags[sort_indices]

print("1. Number of unique tags: ", len(unique_tags))

cvt = CountVectorizer(analyzer = lambda x:x)
ques_tag_map = cvt.fit_transform(tags)
tag_vocab = cvt.vocabulary_

_, tag_count = np.unique(ques_tag_map.nonzero()[1], return_counts=True)
tag_count.sort()

print("\n2. Number of questions associated to a tag: \n", tag_count)

print("\n3. Lowest number of questions associated to a tag: ", np.min(tag_count))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

print(raw_data_old.shape)

<p style='font-size:16px'><b> Inferences (Old data)  </b></p>

<p>

- Total number of points in the dataset = 150000
- Lowest number & % of questions associated to a tag = 116, 0.0073%
    
</p>


<p style='font-size:16px'><b> New data (Tag based Even sampling)  </b></p>


# drop the duplicate rows and create a new dataset

start_time = datetime.datetime.now()

raw_data = raw_data.drop_duplicates(subset = ['Title', 'Body', 'Tags'])

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

tags = [str(x).split(" ") for x in raw_data.Tags]

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# List to store each tagging instance accross all datapoints
tag_list = []

for datapoint in tags:
    for tag in datapoint:
        tag_list.append(tag)

# Calculating the unique tags & their frequency in a descending order from the above tag_list        
unique_tags, tag_count = np.unique(tag_list, return_counts=True)
sort_indices = np.argsort(tag_count)[::-1]
unique_tags = unique_tags[sort_indices]

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Creating a sparse representation of the datapoint/question number and associated tags
cvt = CountVectorizer(analyzer = lambda x:x)
ques_tag_map = cvt.fit_transform(tags)
tag_vocab = cvt.vocabulary_
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Reducing data size for adjusting to PC computational limitations

# Filtering out the datapoints that don't have any of the top 500 tags as the associated tags
num_top_tags = 500
top_tags = unique_tags[:num_top_tags]
tag_inds = [tag_vocab[tag] for tag in top_tags]
relev_ques = ques_tag_map[:, tag_inds]

minq_per_tag = 1000
minq_ind = set()

for i in range(num_top_tags):
    
    ques = relev_ques[:, i].nonzero()[0]
    temp = min(minq_per_tag, len(ques))
    rand_inds = random.sample(range(0, len(ques)), temp)
    minq_ind.update(ques[rand_inds])
    
trimmed_data = raw_data.iloc[list(minq_ind)]
print(trimmed_data.shape)
    
print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# Creating a new column name Tags_shortlist that contains only the top 500 tags from the Tags column
tags_temp = [str(x).split(" ") for x in trimmed_data.Tags]
shortlist_tags = [[item for item in datapoint if item in top_tags] for datapoint in tags_temp]
shortlist_tags = [" ".join(item) for item in shortlist_tags]
trimmed_data['Tags_shortlist'] = shortlist_tags

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

start_time = datetime.datetime.now()

# List to store each tagging instance accross all datapoints
tag_list = []

tags = [str(x).split(" ") for x in trimmed_data.Tags_shortlist]

for datapoint in tags:
    for tag in datapoint:
        tag_list.append(tag)

# Calculating the unique tags & their frequency in a descending order from the above tag_list        
unique_tags, tag_count = np.unique(tag_list, return_counts=True)
sort_indices = np.argsort(tag_count)[::-1]
unique_tags = unique_tags[sort_indices]

print("1. Number of unique tags: ", len(unique_tags))

cvt = CountVectorizer(analyzer = lambda x:x)
ques_tag_map = cvt.fit_transform(tags)
tag_vocab = cvt.vocabulary_

_, tag_count = np.unique(ques_tag_map.nonzero()[1], return_counts=True)
tag_count.sort()

print("\n2. Number of questions associated to a tag: \n", tag_count)

print("\n3. Lowest number of questions associated to a tag: ", np.min(tag_count))

print("\nBlock execution time: ", datetime.datetime.now() - start_time)

<p style='font-size:16px'><b> Inferences (New data)  </b></p>

<p>

- Total number of points in the dataset = 475777
- Lowest number & % of questions associated to a tag = 1105, 0.232%
    
Compared to the old data, the tag with lowest frequency still accounts for 0.232 datapoints compared to 0.0773% in the old dataset. 


</p>


start_time = datetime.datetime.now()

# Save data to csv file
#### Imp: Careful not to overwrite data ###

try:
    trimmed_data.to_csv('dataset 5L/trimmed.csv', mode = 'w', index = False)
except:
    print("File already saved?")

print("\nBlock execution time: ", datetime.datetime.now() - start_time)
